track: fix integer division in speed_check track age factor
track_age and track_max_age are both int64_t, so
(1.0f - track_age / track_max_age) is 1.0 for every track younger than
track_max_age (5 s), which is the only case where the branch runs. The
track bonus (or penalty, for positions against the track) was meant to
fade out with the age of the track and never did. Cast to float.

Factor before -> after: 0.5 s 1.00 -> 0.90, 1 s 1.00 -> 0.80,
2.5 s 1.00 -> 0.50, 4.9 s 1.00 -> 0.02.

A/B on the same live Beast feed (single RTL-SDR receiver at SKCL, 7 h,
232754 global and 1008 local CPR decodes): no speed_check rejections in
either build, all cpr counters identical. The change only matters when
a position is near the speed limit, which this feed did not exercise.

+++ track.c
@@ -597,7 +597,7 @@ static int speed_check(struct aircraft *a, datasource_t source, double lat, doub
 
     if (speed > 10 && track_diff > -1 && a->trackUnreliable < 8) {
         track_bonus = speed * (90.0f - track_diff) / 90.0f;
-        track_bonus *= (surface ? 0.9f : 1.0f) * (1.0f - track_age / track_max_age);
+        track_bonus *= (surface ? 0.9f : 1.0f) * (1.0f - (float) track_age / track_max_age);
         if (a->gs < 10) {
             // don't allow negative "bonus" below 10 knots speed
             track_bonus = fmaxf(0.0f, track_bonus);
